*Slight optimization of eof_menu_track_repair_grid_snap(). *Fixed a bug where toggling lane 6 (black 3 gem) in a GHL track would remove all other gems from the note. *Updated year in copyright string. *Added a "File>Display>Set 3D HOPO image scale size" function allowing the size of non GHL HOPO gems to be changed in the 3D preview. *Added new logic to MIDI import that will automatically resnap notes that lose grid snap due to math errors during the import's timing conversion.

+++ src/song.c
@@ -4220,7 +4220,7 @@ unsigned long eof_get_note_flags(EOF_SONG *sp, unsigned long track, unsigned lon
 	return 0;	//Return error
 }
 
-unsigned short eof_get_note_tflags(EOF_SONG *sp, unsigned long track, unsigned long note)
+unsigned long eof_get_note_tflags(EOF_SONG *sp, unsigned long track, unsigned long note)
 {
 	unsigned long tracknum;
 
@@ -5862,7 +5862,7 @@ void eof_set_note_flags(EOF_SONG *sp, unsigned long track, unsigned long note, u
 	}
 }
 
-void eof_set_note_tflags(EOF_SONG *sp, unsigned long track, unsigned long note, unsigned short tflags)
+void eof_set_note_tflags(EOF_SONG *sp, unsigned long track, unsigned long note, unsigned long tflags)
 {
 // 	eof_log("eof_set_note_flags() entered");
